Remove DispatchThread from unit tests
Fixes: #602

During some of the hangs I've been seeing during test runs I saw a couple of times where AppDomain.Unload wouldn't complete because the DispatchThread was still running. I've spent a large chunk of time trying to figure it out underlying cause without avail. We don't actually make use of the DispatchThread in our tests - so replace with a simplier implementation.

We now use the default JoinableTaskContext which uses the Current thread as the main thread, and uses XUnit's sync context which basically just executes actions on the same thread.
