+++ src/Microsoft.VisualStudio.ProjectSystem.Managed.UnitTests/Mocks/IProjectThreadingServiceFactory.cs
@@ -1,26 +1,16 @@
 ﻿// Copyright (c) Microsoft.  All Rights Reserved.  Licensed under the Apache License, Version 2.0.  See License.txt in the project root for license information.
 
 using System;
-using System.Threading.Tasks;
 
 using Moq;
 
 namespace Microsoft.VisualStudio.ProjectSystem
 {
-    internal static class IProjectThreadingServiceFactory
+    internal static partial class IProjectThreadingServiceFactory
     {
         public static IProjectThreadingService Create()
         {
-            return new IProjectThreadingServiceMock();
-        }
-
-        public static IProjectThreadingService ImplementExecuteSynchronously()
-        {
-            var mock = new Mock<IProjectThreadingService>();
-
-            mock.Setup(s => s.ExecuteSynchronously(It.IsAny<Func<Task>>())).Callback<Func<Task>>(task => task().RunSynchronously());
-
-            return mock.Object;
+            return new ProjectThreadingService();
         }
 
         public static IProjectThreadingService ImplementVerifyOnUIThread(Action action)
